feat: provide `update-peer` command
This runs a workflow of the same name, which updates the bootstrap peer address in the node
registry.

This was necessary because we had removed the bootstrap peer that the `PROD-07` nodes were using.

A small retry loop is also provided for obtaining the workflow run ID. In some cases, the two-second
sleep before trying was not long enough.