feat: provide a `launch-network` command
This command is a bit more complex because the inputs in the YAML document don't map directly to the
inputs in the workflow. However, that's actually what makes this runner much easier to work with.

It also introduces a deployments table in the database, which should make it easier to see what kind
of environment was created.