fix: use generic node size for private node VMs
It was a mistake to use the size of the NAT gateway VMs here. Private nodes always have the same
size as the generic nodes: they use the same Terraform variable.

Also removed some duplicate code by introducing a `reporting` module. There is still some more code
to be moved in here.