feat: create project update for comparison
In addition to an issue being created, we also now create an update for the project. It will have
the same content as the issue. This is useful just to make people more aware that the comparison is
taking place.

The code here was also refactored to reduce duplication by using a helper function to send the
requests to Linear and handle errors.