feat: provide `deployments linear` cmd
The Linear API will be used to create an issue and a project update related to the deployment.
Almost all the time, the deployment is related to some kind of test, which the user will get to
choose. That choice then forms the some of the basis for the title of the issue that gets created.