chore: make network id mandatory for client deployment
It is now mandatory to supply a network ID for a client deployment. It was changed to a requirement
on `testnet-deploy` just to make things unambiguous.

There is another change here for renaming the `network-name` input to `deployment-name`. This is
because we are not deploying a network for the client deployment.