feat: provide `releases new` cmd
Creates a new release by adding a project on Linear and populating it with several initial issues.

There's also a `releases breaking` command to examine a list of PRs to find which of them have
breaking changes. This is useful when putting the release together so that we can verify if the
crate versions have been bumped correctly.