feat: provide new `releases` commands
The existing `releases new` command was broken into two commands: `new` and `new-rc`.

I have decided now to have two different types of projects: one for releases and the other for RCs.

This is because there can be many RCs before a release, with the final RC being promoted to the
release.